This PR addresses random test failures observed in GitHub Actions CI. The failures were caused by timing-sensitive tests that behaved inconsistently under CI load and across different operating systems.

**Changes:**
- **cmd/root_test.go**: Made timing assertions more resilient to CI variability
  - `TestAwaitDockerClient`: Widened timing assertion range from `>=1s <2s` to `>=900ms <3s`
  - `TestUpdateLockSerialization`: Increased simulation sleep from 10ms to 100ms
  - `TestConcurrentScheduledAndAPIUpdate`: Increased sleeps from 50ms/100ms to 100ms/200ms

- **internal/actions/update_test.go**: Fixed safeguard delay test
  - Changed `PullFailureDelay` from 1ms to 10ms to prevent timing failures on slower systems